// api-new/functions/src/models/enums.ts
export enum RESPONSE_STATUS {
    SUCCESS = "success",
    FAILURE = "failure",
    ERROR = "error"
}

export enum TRANS_TYPE {
    EXPENSE = "expense",
    INCOME = "income",
    OTHER = "other",
    PHONE = "phone",
    INTERNET = "internet",
    RESTAURANT = "restaurant",
    MARKET = "market",
    ENTERTAINMENT = "entertainment",
    SOFTWARE = "software",
    FURNITURE = "furniture",
    UTILITY = "utility",
    PET = "pet",
    CLOTHING = "clothing"
}